requirements.txt
arklex==0.0.9rc2
langchain>=0.1.0
openai>=1.12.0
python-dotenv>=1.0.0
langgraph>=0.0.10
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0.1
pydantic>=2.5.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiohttp>=3.9.1
tqdm>=4.66.1
colorama>=0.4.6
webdriver-manager>=4.0.1